Update README.md
Launch argument -no_render_on_input_thread has been made the default for players with 8 or more cores on their CPU. This improves performance stability and input latency, especially when paired with high-polling rate mice or controllers. It can still be enabled manually for players with fewer CPU cores, and it can be disabled with -render_on_input_thread.